Analytics fix
